datastore 관련 클래스 datastore 모듈로 이동
data 모듈 내에 datastore 의존성 제거
