+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/di/DataSourceModule.kt
@@ -1,17 +1,9 @@
 package com.nexters.bandalart.android.core.data.di
 
 import com.nexters.bandalart.android.core.data.datasource.BandalartRemoteDataSource
-import com.nexters.bandalart.android.core.data.datasource.CompletedBandalartKeyDataSource
-import com.nexters.bandalart.android.core.data.datasource.GuestLoginLocalDataSource
 import com.nexters.bandalart.android.core.data.datasource.GuestLoginRemoteDataSource
-import com.nexters.bandalart.android.core.data.datasource.OnboardingDataSource
-import com.nexters.bandalart.android.core.data.datasource.RecentBandalartKeyDataSource
-import com.nexters.bandalart.android.core.data.local.datasource.CompletedBandalartKeyDataStoreImpl
-import com.nexters.bandalart.android.core.data.local.datasource.GuestLoginLocalDataSourceImpl
-import com.nexters.bandalart.android.core.data.local.datasource.OnboardingDataSourceImpl
-import com.nexters.bandalart.android.core.data.local.datasource.RecentBandalartKeyDataSourceImpl
-import com.nexters.bandalart.android.core.data.remote.datasource.BandalartRemoteDataSourceImpl
-import com.nexters.bandalart.android.core.data.remote.datasource.GuestLoginRemoteDataSourceImpl
+import com.nexters.bandalart.android.core.data.datasource.BandalartRemoteDataSourceImpl
+import com.nexters.bandalart.android.core.data.datasource.GuestLoginRemoteDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -24,37 +16,13 @@ internal abstract class DataSourceModule {
 
   @Binds
   @Singleton
-  abstract fun bindGuestLoginLocalDataSource(
-    guestLoginLocalDataSourceImpl: GuestLoginLocalDataSourceImpl,
-  ): GuestLoginLocalDataSource
-
-  @Binds
-  @Singleton
   abstract fun bindGuestLoginRemoteDataSource(
     guestLoginRemoteDataSourceImpl: GuestLoginRemoteDataSourceImpl,
   ): GuestLoginRemoteDataSource
 
   @Binds
   @Singleton
-  abstract fun bindRecentBandalartKeyDataSource(
-    recentBandalartKeyDataSourceImpl: RecentBandalartKeyDataSourceImpl,
-  ): RecentBandalartKeyDataSource
-
-  @Binds
-  @Singleton
   abstract fun bindBandalartRemoteDataSource(
     bandalartRemoteDataSourceImpl: BandalartRemoteDataSourceImpl,
   ): BandalartRemoteDataSource
-
-  @Binds
-  @Singleton
-  abstract fun bindCompletedBandalartKeyDataSource(
-    completedBandalartKeyDataStoreImpl: CompletedBandalartKeyDataStoreImpl,
-  ): CompletedBandalartKeyDataSource
-
-  @Binds
-  @Singleton
-  abstract fun bindOnboardingDataSource(
-    onboardingDataSourceImpl: OnboardingDataSourceImpl,
-  ): OnboardingDataSource
 }
