+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/repository/GuestLoginTokenRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.nexters.bandalart.android.core.data.repository
 
-import com.nexters.bandalart.android.core.data.datasource.GuestLoginLocalDataSource
+import com.nexters.bandalart.android.core.datastore.datasource.GuestLoginLocalDataSource
 import com.nexters.bandalart.android.core.data.datasource.GuestLoginRemoteDataSource
 import com.nexters.bandalart.android.core.data.mapper.toEntity
 import com.nexters.bandalart.android.core.domain.entity.GuestLoginTokenEntity
