+++ core/datastore/src/main/kotlin/com/nexters/bandalart/android/core/datastore/datasource/OnboardingDataSource.kt
@@ -1,4 +1,4 @@
-package com.nexters.bandalart.android.core.data.datasource
+package com.nexters.bandalart.android.core.datastore.datasource
 
 interface OnboardingDataSource {
   suspend fun setOnboardingCompletedStatus(flag: Boolean)
