+++ core/datastore/src/main/kotlin/com/nexters/bandalart/android/core/datastore/datasource/RecentBandalartKeyDataSource.kt
@@ -1,4 +1,4 @@
-package com.nexters.bandalart.android.core.data.datasource
+package com.nexters.bandalart.android.core.datastore.datasource
 
 interface RecentBandalartKeyDataSource {
   suspend fun setRecentBandalartKey(recentBandalartKey: String)
